rename StudentDetails -> StudentCV
